Created MovieQuizPresenter. Moved the method convert to MovieQuizPresenter. Used DRY to eliminate data duplication.
