+++ MovieQuiz/Models/QuizResultsViewModel.swift
@@ -12,4 +12,3 @@ struct QuizResultsViewModel {
     let text: String
     let buttonText: String
 }
-// Предлагаем вам вывести в сообщении алерта статистику игр пользователя с помощью созданного класса  StatisticServiceImplementation
