+++ MovieQuiz/Services/MoviesLoader.swift
@@ -20,7 +20,7 @@ struct MoviesLoader: MoviesLoading {
     
     // MARK: - URL
     private var mostPopularMoviesUrl: URL {
-        guard let url = URL(string: "https://imdb-api.com/en/API/Top250Movies/k_kiwxbi4y") else {
+        guard let url = URL(string: "https://imdb-api.com/en/API/Top250Movies/k_zcuw1ytf") else {
             preconditionFailure("Unable to construct mostPopularMoviesUrl")
         }
         return url
@@ -42,3 +42,4 @@ struct MoviesLoader: MoviesLoading {
         }
     }
 }
+
